+++ main.py
@@ -49,6 +49,7 @@ def level1():  # main function of our game
     fixed_2 = False
     fixed_3 = False
     fixed_4 = True
+    nxt = False
     clock = pygame.time.Clock()
 
     # main loop of our game
@@ -60,8 +61,15 @@ def level1():  # main function of our game
             if event.type == pygame.QUIT:
                 run = False
 
+
         # getting the keys pressed
         keys_pressed = pygame.key.get_pressed()
+
+        if keys_pressed[pygame.K_q] and keys_pressed[pygame.K_ESCAPE]:
+            main_menu()
+
+        if fixed_1 and fixed_2 and fixed_3 and fixed_4:
+            nxt = True
 
         # changing the coordinates using the pygame buttons tracker
         character_rectangle.x, character_rectangle.y = movement(VELOCITY, character_rectangle, keys_pressed)
@@ -72,8 +80,10 @@ def level1():  # main function of our game
         text, color, fixed_1, fixed_2, fixed_3, mistakes = interact1(character_rectangle, keys_pressed, fixed_1, fixed_2, fixed_3, level, mistakes)
 
         # refreshing the picture
-        refresh(character_rectangle, CONSTANTS.CHARACTER, text, color, keys_pressed, level, fixed_1, fixed_2, fixed_3, fixed_4, mistakes)
+        refresh(character_rectangle, CONSTANTS.CHARACTER, text, color, keys_pressed, level, fixed_1, fixed_2, fixed_3, fixed_4, mistakes, nxt)
 
+        if nxt == True and character_rectangle.colliderect(DOOR_RECT):
+            level5()
     pygame.quit()
 
 def level5():
@@ -85,6 +95,9 @@ def level5():
     fixed_2 = False
     fixed_3 = False
     fixed_4 = False
+    nxt = False
+    text = ''
+    color = ''
     clock = pygame.time.Clock()
 
     # main loop of our game
@@ -99,6 +112,12 @@ def level5():
         # getting the keys pressed
         keys_pressed = pygame.key.get_pressed()
 
+        if keys_pressed[pygame.K_q] and keys_pressed[pygame.K_ESCAPE]:
+            main_menu()
+
+        if fixed_1 and fixed_2 and fixed_3 and fixed_4:
+            nxt = True
+
         # changing the coordinates using the pygame buttons tracker
         character_rectangle.x, character_rectangle.y = movement(VELOCITY, character_rectangle, keys_pressed)
         # changing the model according to where character moves
@@ -108,7 +127,8 @@ def level5():
         text, color, fixed_1, fixed_2, fixed_3, mistakes = interact5(character_rectangle, keys_pressed, fixed_1, fixed_2, fixed_3, fixed_4, level, mistakes)
 
         # refreshing the picture
-        refresh(character_rectangle, CONSTANTS.CHARACTER, text, color, keys_pressed, level, fixed_1, fixed_2, fixed_3, fixed_4, mistakes)
+        refresh(character_rectangle, CONSTANTS.CHARACTER, text, color, keys_pressed, level, fixed_1, fixed_2, fixed_3, fixed_4, mistakes, nxt)
+        print(character_rectangle.x, character_rectangle.y)
 
     pygame.quit()
 
@@ -116,4 +136,4 @@ def level5():
 
 # calling the main function
 if __name__ == "__main__":
-    level5()
+    level1()
